chore: decompose monolothic CLI router
