+++ src/obsidian_ai_tools.egg-info/SOURCES.txt
@@ -33,6 +33,8 @@ src/obsidian_ai_tools.egg-info/dependency_links.txt
 src/obsidian_ai_tools.egg-info/entry_points.txt
 src/obsidian_ai_tools.egg-info/requires.txt
 src/obsidian_ai_tools.egg-info/top_level.txt
+src/obsidian_ai_tools/commands/__init__.py
+src/obsidian_ai_tools/commands/notes.py
 src/obsidian_ai_tools/providers/__init__.py
 src/obsidian_ai_tools/providers/base.py
 src/obsidian_ai_tools/providers/factory.py
@@ -59,11 +61,13 @@ tests/test_integration.py
 tests/test_llm.py
 tests/test_logging.py
 tests/test_models.py
+tests/test_observability.py
 tests/test_obsidian.py
 tests/test_overview.py
 tests/test_preview.py
 tests/test_refresh.py
 tests/test_search.py
 tests/test_tag_hygiene.py
+tests/test_transcript_validation.py
 tests/test_wikilinks.py
 tests/test_youtube.py
