+++ src/obsidian_ai_tools/commands/__init__.py
@@ -1,0 +1,1 @@
+"""CLI command modules — one file per command group."""
